POST data from frontend to backend
Things aren't perfect. Need to figure out how to get the form data into
a nicer JSON format to fit the expected format.

Still having that issue with this.state where it doesn't update
immediately, but that shouldn't be a big deal since the page will
refresh after the response is received from the backend