Creating event properly with multiple attendees
There is a minor issue where "No access, refresh token or API key is
set." is output before the event is created, but creation is going fine.
I think it's just because there is no access token.

Can be adapted rather easily to support attendees from frontend

NOTE: THIS WAS ASYNC/AWAIT HELL PT 2. Async/await in a for-loop SUCKS.